Update patch set 3
Patch Set 3: Code-Review-1

(6 comments)

Im confused... is this for a new .git repo? the name of the file should match that name. 

If this is for a subfolder in cps, then the projects should still be "project:cps"

Patch-set: 3
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1