Revert "Update sonar scanner default to latest..."
This reverts commit f85c92920e00ce776ed25e70f1cd87678f2f02ab.
The commit did not have the intended effect. It was meant to replace
an outdated version of the sonar plugin, but this latest version is
not available for mvn builds. The former version is the latest
available via mvn, so we should just revert this commit. The fix for
this problem will ultimately be to update the node.js version on the
builder prior to running the scan.

Change-Id: I4acf72e50c3f54b7b49a71833bdd7648de609802
Issue-ID: CIMAN-33
Signed-off-by: Eric Ball <eball@linuxfoundation.org>