Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

The project and project-name are a little mismatched. The good thing is that project is set to the correct value but do we want to make these two match ? Also.. the name of the file, should't that be integration-simulators-nf-simulator-avcn-manager.yaml?

Patch-set: 1
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1