Set builder name and type to "openstack"
Based on the input provided on the Gihub packer issue
hard-coded for our parser to be able to decode the rest of the config
properly." This would require to hardcode the builder name and type
in the templates.

This change requires an update in global-jjb v0.60.3
and common-packer v0.7.4.

Change-Id: I506959b8c1999619877432ebba7a6947ce0b0c2c
LF-JIRA: https://jira.linuxfoundation.org/browse/RELENG-3448
Issue-ID: CIMAN-33
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>