Add script to update nodejs for Sonarcloud scans
Some projects have been seeing failures in Sonarcloud scans, as they
are no longer compatible with Node <v10. We were unable to find the
requirement that is pulling Node in, and it is not installed by
default on the CI agents.

By installing an up-to-date version before running any other tasks,
the issue is fixed and the Sonarcloud scans do not run into any
issues. All affected jobs have been updated.

Change-Id: I01a578b37300b6dd6e6fd235218978a19518de99
Issue-ID: CIMAN-33
Signed-off-by: Eric Ball <eball@linuxfoundation.org>