Update patch set 1
Patch Set 1:

we're still using npm.public, but need TWO DIFFERENT URLs.  The first is found in our pom (which I'm guessing is similar to these yaml files) - we had to change it to this:
https://nexus3.onap.org/repository/npm.public/npm/-/

that works in conjunction with the second, which appears in the npmrc:
https://nexus3.onap.org/repository/npm.public/

Note: the first URL is needed to download the npm.tar.gz, while the second is needed to download all of its components.

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>