Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> Jess - I do not see new gerrit-docker-merge-maintainance job being added - any clue on why?
> 
> Only see dcaegen2-services-pmsh-docker-merge-master (https://jenkins.onap.org/view/dcaegen2-services-pmsh/) which is same as before and running on master

You don't have a docker job under the ID "gerrit-docker-merge-maintainance" .. are you adding one?
I think you wanted to use gerrit-docker-merge? with the honolulu stream is it? If you are planning to use the docker merge job from global-jjb, the ID should be gerrit-docker-merge

If so, please make sure to verify this in the sandbox.

Patch-set: 8