Update patch set 1
Patch Set 1: Code-Review-1

Can you please test this in the sandbox? I am not sure if the parameter gets passed since this variable is not defined in the original global-jjb template. If it does not work, then we will need to add it in global-jjb first.

Patch-set: 1
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1