Skip test in staging jobs
Tests fail in the staging jobs due to differences between the Ubuntu
environment used in the verify and merge jobs and the Centos environment
used in the staging jobs. Specifically, the "keytool" utility that comes
in Java is not in the path in CentOS.

As the tests must pass in the Ubuntu environment ver verification and
merge, there is no need to run the tests at staging. THis review
disables tests in staging jobs.

Issue-ID: POLICY-3587
Change-Id: I357e15ce908491a1849f64d33a7c9df47471de39
Signed-off-by: liamfallon <liam.fallon@est.tech>