Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I didn't think the CLM tool could handle branches. Perhaps this changed but long ago I had multiple branches setup but it would only have one report. The report would be overwritten depending on the last job that ran.

This is all work to deal with the log4j issue.

Yes, I was surprised to see that too, but all the PTLs got an email telling us to put this change in so that the CLM jobs can run on branches other than Master. According to the email, if you follow the "view report" link in the Jenkins link of the CLM job on a branch, then you'll get the report for that branch.  

I'll forward the email to all committers.

Patch-set: 1