Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> See following error on sandbox run- 
> 
> 17:40:04 [dcaegen2-collectors-snmptrap-tox-verify-master] $ /bin/sh -xe /tmp/jenkins5467816286450901938.sh
> 17:40:04 + ../../shell/onap-gerrit-review-wrapper.sh
> 17:40:04 /tmp/jenkins5467816286450901938.sh: 2: /tmp/jenkins5467816286450901938.sh: ../../shell/onap-gerrit-review-wrapper.sh: not found
> 17:40:04 Build step 'Execute shell' marked build as failure
> 
> Should the script be moved into sandbox env for testing this change?

So.. if you look at your configuration, notice that script box is not getting filled with the contents of the script.. but rather the location of the script. 

https://jenkins.onap.org/sandbox/job/dcaegen2-collectors-snmptrap-tox-verify-master/configure

You need to have this box filled with the actual contents of the script.. 
i think you need to use "!include-raw-escape:" 
You can see how this is used in other yaml files in ci-management.

Patch-set: 3
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>