Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does this change needs to be proposed in https://github.com/lfit/releng-global-jjb/blob/master/jjb/lf-rtdv3-jobs.yaml?

Here it overrides without modifying the global jjbs which are out of ONAP.
Yes I would suggest to comply with standards in global jjbs but my focus is to clean ONAP jjb. And, in a long, the ONAP doc job must be fully rewritten and free from global jjbs.

This can be safely merged.

Patch-set: 1