Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> The SBOM generator runs by default in the $WORKSPACE. 
> 
> Please run this locally in your machine so that you are able to see the error message with the dependencies.

I verified with local run - this update is more of workaround to address the SBOM executable issues (https://github.com/opensbom-generator/spdx-sbom-generator/issues/227, emailed on this on Friday). 

We need to change LF wrapper update to use the -p path as override to copy the executable itself. This specific change is not required if LF wrapper can use $WORKSPACE as default to copy the binary (when -p path override is not specified).

Patch-set: 1