Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Sorry,, I am little bit confused.. so,, you were able to test this locally and you are no longer getting the transitive dependencies issues. 
> I tested your change and it is the same result:
> https://jenkins.onap.org/job/dcaegen2-collectors-ves-maven-stage-master/1015/
> 
> I overwrote the job to use -p $WORKSPACE but it didn't had an effect since this job is already using the $WORKSPACE path as default:

Hey Jess - To avoid the transitive dependency error, we need to copy the spdx-sbom-generator binary to -p specified path before executing it. This is a required as the path parameter is not being read by SBOM binary currently; copying the binary to target/required path gets pass the transitive dependency error. 

This workaround may be required under SBOM generator community fixes the issue.

Majority of ONAP project would have $WORKSPACE as root, but that is not true for all cases (e.g dcaegen2/platform or dcaegen2/services) where we can override the -p to required path for those subcomponent. 

So once the LF wrapper script -“sbom-generator.sh” is updated to copy the binary to specified path (and switch to this directory) before invoking it - we should be good.

Patch-set: 1