Update patch set 2
Patch Set 2:

Tried doing git submodule update --init, and still see changes to global-jjb in this review, which should not be there.  It'll be simpler just to abandon this review and submit a new one ... probably made the mistake of using "git add ." instead of just adding the jjb changes.

Patch-set: 2