Revert global-jjb to v0.92.7
v0.92.8 introduces an issue with some mvn jobs. This was change was
made due to a reported bug, and fixed the job in question, but seems
to have a greater negative impact on other jobs. Because it is a
release blocker, we will roll back for now, and then update once a
more universal fix is in place.

Change-Id: Ie20cba8a30e95fec7d17643fffb419e31f2601b0
Issue-ID: CIMAN-33
Signed-off-by: Eric Ball <eball@linuxfoundation.org>